third_party/blink/web_tests/webaudio/Analyser/realtimeanalyser-zero.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test getFloatFrequencyData With Zero Inputs
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
  </head>
  <body>
    <script>
      const sampleRate = 48000;

      // Render enough data to run the test.
      const renderFrames = 2 * 1024;
      const renderDuration = renderFrames / sampleRate;

      // Test that getFloatFrequencyData returns -Infinity when the input is
      // all-zeroes.
      promise_test(async () => {
        const context = new OfflineAudioContext(1, renderFrames, sampleRate);

        // Constant source of 0's.
        const source = new ConstantSourceNode(context, {offset: 0});

        // Create analyser and use some non-default minDecibels value.
        const analyser = new AnalyserNode(context, {minDecibels: -123});

        source.connect(analyser).connect(context.destination);

        source.start();

        // Suspend after some number of frames and grab the float frequency
        // data.
        context.suspend(1024 / sampleRate).then(() => {
          const f = new Float32Array(analyser.frequencyBinCount);
          analyser.getFloatFrequencyData(f);

          assert_true(
              f.every(x => x === -Infinity),
              'getFloatFrequencyData() should contain only -Infinity');
        }).then(() => context.resume());

        await context.startRendering();
      }, 'zero input');
    </script>
  </body>
</html>
